fix(reference-pack): a clippy sweep, a doc link, and a scratch-file race (MJXOFF-207)
Three things the full sweep found that the per-suite runs did not.

* **Clippy**, across the new crate and the two it touches: an empty line inside a doc comment (the
  block was prose about the module rather than about the item below it, so it is a `//` comment
  now), four `u32::from(u32)` conversions, a `chunks_exact(3)` that wants `as_chunks::<3>`, and an
  assertion on constants that wants a `const` block.
* **A redundant explicit doc link**, which `RUSTDOCFLAGS='-D warnings'` refuses.
* **A race between three cases sharing one scratch PDF.** Cargo runs a binary's cases on several
  threads, and three of `the_plumbing_is_proved_and_not_the_fidelity.rs`'s cases wanted page one of
  the same export. One rasterised the file another was halfway through writing, and `pdftoppm`
  reported *"Document stream is empty"* — a flake that reads exactly like an exporter defect. The
  scratch name now carries the case's own name, and the suite was run five times over to confirm.

+++ crates/mjx-pptx/tests/preset_adjustments.rs
@@ -242,7 +242,7 @@ fn every_adjustment_of_every_preset_is_writable_by_wire_name() {
         for spec in adjustments_of(*preset) {
             // Off the default by a value no default is, so a writer that silently did nothing would
             // read back the default and fail. `saturating_add` because a default may be extreme.
-            let target = i32::from(spec.default).saturating_add(1_234);
+            let target = spec.default.saturating_add(1_234);
             *distinct_targets.entry(target).or_default() += 1;
             deck.set_shape_adjustments(0, shape, &[(spec.wire_name, target)])
                 .expect("the adjustment is written");
